fix: break installation cycle in database context binding
